Version 1.2 Proyecto GameStation
se logro cargar el sprite del carro y moverlo

+++ src/main/java/org/GameStation/AlmacenamientoPuntaje.java
@@ -1,10 +1,8 @@
 package org.GameStation;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 
-public class AlmacenamientoPuntaje {//almacenamiento temporal, se evalua el uso de gson
+public class AlmacenamientoPuntaje {
 
     public void crearArchivo(String ruta) {
         File archivo = new File(ruta);
@@ -28,7 +26,33 @@ public class AlmacenamientoPuntaje {//almacenamiento temporal, se evalua el uso 
 
     public void eliminarArchivo(String ruta) {
         File archivo = new File(ruta);
-        archivo.delete();
+        if (archivo.exists()) {
+            archivo.delete();
+        } else {
+            System.out.println("El archivo no existe: " + ruta);
+        }
+    }
+
+    public String leerArchivo(String ruta) {
+        File archivo = new File(ruta);
+        StringBuilder contenido = new StringBuilder();
+
+        if (archivo.exists()) {
+            try {
+                BufferedReader reader = new BufferedReader(new FileReader(archivo));
+                String linea;
+
+                while ((linea = reader.readLine()) != null) {
+                    contenido.append(linea);
+                }
+
+                reader.close();
+            } catch (IOException e) {
+                System.out.println("Error al leer el archivo: " + e.getMessage());
+            }
+        }
+
+        return contenido.toString();
     }
 
 }
